refactor(arch): move database file from setting to common
