+++ pkg/common/database.go
@@ -1,9 +1,10 @@
-package setting
+package common
 
 import (
 	"fmt"
 
 	"github.com/arashi87/gin-template/pkg/model"
+	"github.com/arashi87/gin-template/pkg/setting"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
@@ -16,7 +17,8 @@ var DB *gorm.DB
 */
 func InitDatabase() *gorm.DB {
 	dsn := fmt.Sprintf("host=%s user=%s password=%s dbname=%s port=%s sslmode=disable TimeZone=%s",
-		CONFIG.DBHost, CONFIG.DBUsername, CONFIG.DBPassword, CONFIG.DBName, CONFIG.DBPort, CONFIG.DBTimezone)
+		setting.CONFIG.DBHost, setting.CONFIG.DBUsername, setting.CONFIG.DBPassword,
+		setting.CONFIG.DBName, setting.CONFIG.DBPort, setting.CONFIG.DBTimezone)
 
 	// connect to database
 	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
